requirements.txt
streamlit
google-generativeai>=0.5.0
python-dotenv
pdf2image